refactor: 유저 관련 enums를 user/domain 하위로 이동
