+++ src/main/java/bitnagil/bitnagil_backend/auth/kakao/domain/CustomOAuth2User.java
@@ -6,8 +6,7 @@ import java.util.Map;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
 
-import bitnagil.bitnagil_backend.enums.Role;
-import bitnagil.bitnagil_backend.global.entity.HistoryPk;
+import bitnagil.bitnagil_backend.user.domain.enums.Role;
 import lombok.Getter;
 
 /**
