+++ src/main/java/bitnagil/bitnagil_backend/auth/kakao/domain/OAuth2Attribute.java
@@ -3,11 +3,11 @@ package bitnagil.bitnagil_backend.auth.kakao.domain;
 import java.util.HashMap;
 import java.util.Map;
 
-import bitnagil.bitnagil_backend.enums.SocialType;
+import bitnagil.bitnagil_backend.user.domain.enums.SocialType;
 import bitnagil.bitnagil_backend.global.errorcode.ErrorCode;
 import bitnagil.bitnagil_backend.global.exception.CustomException;
 import bitnagil.bitnagil_backend.user.domain.User;
-import bitnagil.bitnagil_backend.enums.Role;
+import bitnagil.bitnagil_backend.user.domain.enums.Role;
 import lombok.Builder;
 import lombok.Getter;
