--- src/main/java/bitnagil/bitnagil_backend/enums/SocialType.java
@@ -1,6 +1,0 @@
-package bitnagil.bitnagil_backend.enums;
-
-public enum SocialType {
-
-    KAKAO, APPLE
-}
