+++ src/main/java/bitnagil/bitnagil_backend/user/domain/enums/SocialType.java
@@ -1,0 +1,6 @@
+package bitnagil.bitnagil_backend.user.domain.enums;
+
+public enum SocialType {
+
+    KAKAO, APPLE
+}
