+++ src/main/java/bitnagil/bitnagil_backend/user/repository/UserRepository.java
@@ -1,14 +1,11 @@
 package bitnagil.bitnagil_backend.user.repository;
 
-import java.time.LocalDateTime;
 import java.util.Optional;
-import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import bitnagil.bitnagil_backend.enums.SocialType;
-import bitnagil.bitnagil_backend.global.entity.HistoryPk;
+import bitnagil.bitnagil_backend.user.domain.enums.SocialType;
 import bitnagil.bitnagil_backend.user.domain.User;
 
 @Repository
